move GatewayServiceFutureStubProvider
