+++ hypertrace-core-graphql-gateway-service-utils/build.gradle.kts
@@ -10,6 +10,7 @@ dependencies {
   api(project(":hypertrace-core-graphql-attribute-store"))
   api("io.reactivex.rxjava3:rxjava")
   api(project(":hypertrace-core-graphql-common-schema"))
+  implementation(project(":hypertrace-core-graphql-grpc-utils"))
 
   testImplementation("org.junit.jupiter:junit-jupiter")
   testImplementation("org.mockito:mockito-core")
